backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.1
python-multipart==0.0.6
numpy==1.24.3